fix: seperator to separator subcomponent
