--- packages/Card/src/components/Seperator/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Seperator";
