+++ packages/Card/stories/examples/DeluxeCard.js
@@ -13,12 +13,12 @@ export default () => {
         <Card.Content>
           <Card.Title>Asset type name</Card.Title>
           <Input />
-          <Card.Seperator size="small" />
+          <Card.Separator size="small" />
           <Checkbox>Use a display name</Checkbox>
-          <Card.Seperator size="small" />
+          <Card.Separator size="small" />
           <Card.Title>Description (optional)</Card.Title>
           <Textarea placeholder="Kanye is the only one who can edit this.." />
-          <Card.Seperator size="small" />
+          <Card.Separator size="small" />
           <div style={{ width: "100%", display: "flex", flexDirection: "row", justifyContent: "space-between" }}>
             <div style={{ width: "50%", display: "flex", justifyContent: "space-between" }}>
               <Button isDisabled="true">Save</Button>
@@ -28,7 +28,7 @@ export default () => {
             </div>
             <Button kind="minor">Archive</Button>
           </div>
-          <Card.Seperator size="small" />
+          <Card.Separator size="small" />
           <Card.Metadata>Created by Cory McBain</Card.Metadata>
           <Card.Metadata>Last updated by Charles Bradley</Card.Metadata>
           <Button kind="link">See full activity log</Button>
