+++ packages/Card/stories/examples/MultipleCard.js
@@ -102,9 +102,9 @@ export default () => {
                 <Card.Content>
                   <Card.Title>{props.title}</Card.Title>
                   <Card.Metadata>{props.meta1}</Card.Metadata>
-                  <Card.Seperator size="small" />
+                  <Card.Separator size="small" />
                   <Card.Text isTruncate>{props.text}</Card.Text>
-                  <Card.Seperator size="small" />
+                  <Card.Separator size="small" />
                   <Card.Metadata>{props.meta2}</Card.Metadata>
                   <Card.Metadata>{props.meta3}</Card.Metadata>
                 </Card.Content>
